resume stepper animation
